IA-DEVELOPER-N2/U06-Integracion-IA-Aplicaciones-Reales/02-Laboratorios/Lab-6.1-Image-Classification/requirements.txt
# Laboratorio 6.1 - Clasificación de Imágenes con FastAPI y React
# Dependencias Backend (Python)
fastapi==2026.3.0
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
pillow==10.0.0
numpy==1.24.3
prometheus-client==0.17.1
redis==4.5.4
pydantic==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.0
httpx==0.25.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
black==23.9.0
isort==5.12.0
flake8==6.1.0
